third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/disable-breakpoints-expected.txt

Tests disabling breakpoints.


Running: testSetBreakpointPauseResumeThenDisable
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (disable-breakpoints.js:13)
Breakpoint sidebar pane 
disable-breakpoints.js:13checked          var x = Math.sqrt(10);
Script execution resumed.
Test function finished.
Disabling breakpoints...
Running test function again...
Set timer for test function.
Test function finished.

Running: testEnableBreakpointsAgain
Enabling breakpoints...
Running test function...
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (disable-breakpoints.js:13)
Breakpoint sidebar pane 
disable-breakpoints.js:13checked          var x = Math.sqrt(10);
Script execution resumed.
Test function finished.
Breakpoint sidebar pane 
disable-breakpoints.js:13checked          var x = Math.sqrt(10);
Breakpoints removed.
Breakpoint sidebar pane 
No breakpoints
